+++ northwing/View/ProductoView.cs
@@ -32,16 +32,34 @@ namespace northwing
         {
             //botón consultar
 
+            try
+            {
+                decimal id = decimal.Parse(this.textBoxproductID.Text); 
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show("Dato introducido incorrecto,por favor introduzca un valor numérico");
+                borrarDatos();
+                return; 
+            }
+
             ds = productoController.consultaTablaProductos(this.textBoxproductID.Text);
-            
+
             if (ds.Products.Rows.Count == 0)
             {
                 MessageBox.Show("El producto no existe,si desea añadir un producto rellene los datos y pulse el botón AÑADIR PRODUCTO");
                 borrarDatos();
                 this.textBoxproductID.Enabled = false; //método para impedir acceso a usuario
-                this.btAlta.Visible = true;
-                this.lbmensajemodificar.Visible = false;
-               
+                this.btañadirproducto.Visible = true;
+                this.panel1.Visible = true;
+                this.lbinstruccion.Visible = true;
+                this.lbsubtitulo.Visible = true;
+                this.lbmensajemodificar.Visible =false;
+
+                this.textBoxproductName.Enabled =true ;
+                this.textBoxsupplierID.Enabled = true;
+                this.cbCategorias.Enabled = true;
+                this.textBoxcantidad.Enabled = true;
             }
             else
             {    
@@ -53,9 +71,18 @@ namespace northwing
 
                 this.bteliminarproducto.Visible=true;
                 this.btModificar.Visible=true;
-                this.lbmensajemodificar.Visible = true;
                 this.btCancelar.Visible = true;
-                MessageBox.Show("Si desea realizar otra consulta pulse BORRAR DATOS");
+                this.panel1.Visible = true;
+                this.lbinstruccion.Visible = false;
+                this.lbsubtitulo.Visible=true;
+                this.lbmensajemodificar.Visible = true;
+
+                this.textBoxproductName.Enabled = false;
+                this.textBoxsupplierID.Enabled = false;
+                this.cbCategorias.Enabled = false;
+                this.textBoxcantidad.Enabled = false;
+
+                MessageBox.Show("Si desea realizar otra consulta pulse BORRAR DATOS"); 
             }
             ds.Products.Clear(); 
             
@@ -82,10 +109,11 @@ namespace northwing
                     {
                         MessageBox.Show("Producto añadido con éxito");
                         borrarDatos();
-                        this.btAlta.Visible=false;
+                        this.btañadirproducto.Visible=false;
                         this.btModificar.Visible = false;
                         this.bteliminarproducto.Visible = false;
                         this.btCancelar.Visible = false;
+                        
                     }
                    
                 }
@@ -114,28 +142,42 @@ namespace northwing
 
         private void Products_Load(object sender, EventArgs e)
         {
-            this.btAlta.Visible = false;
+            this.btañadirproducto.Visible = false;
             this.btCancelar.Visible = false;
             this.btModificar.Visible = false;
-            this.bteliminarproducto.Visible = false;    
+            this.bteliminarproducto.Visible = false;
+            this.textBoxcategoryID.Enabled = false;
+            this.panel1.Visible = false;
+            if (ds.Categories.Rows.Count == 0)
+            {
+                string query = "SELECT CategoryID, CategoryName FROM Categories;";
 
-            string query = "SELECT CategoryID, CategoryName FROM Categories;";
+                SqlCommand command = new SqlCommand(query, connection);
+                SqlDataAdapter adapter = new SqlDataAdapter(command);
 
-            SqlCommand command = new SqlCommand(query, connection);
-            SqlDataAdapter adapter = new SqlDataAdapter(command);
+                DataRow row = ds.Categories.NewRow();
+                row[0] = 0;
+                row[1] = "Elige categoria";
+                ds.Categories.Rows.InsertAt(row, 0);
+                adapter.Fill(ds.Categories);
 
-            DataRow row = ds.Categories.NewRow();
-            row[0] = 0;
-            row[1] = "Elige categoria";
-            ds.Categories.Rows.InsertAt(row, 0);
-            adapter.Fill(ds.Categories);
+                this.cbCategorias.DataSource = ds.Categories;
 
-            this.cbCategorias.DataSource = ds.Categories;
+                this.cbCategorias.DisplayMember = "CategoryName"; //lo que ve usuario
+                this.cbCategorias.ValueMember = "CategoryID"; //valor que guarda el programa
+                this.cbCategorias.SelectedIndex = 0;
+                this.textBoxcategoryID.Text = string.Empty;
+            }
+            else
+            {
 
-            this.cbCategorias.DisplayMember = "CategoryName"; //lo que ve usuario
-            this.cbCategorias.ValueMember = "CategoryID"; //valor que guarda el programa
-            this.cbCategorias.SelectedIndex = 0;
-            this.textBoxcategoryID.Text = string.Empty;
+                this.cbCategorias.DataSource = ds.Categories;
+
+                this.cbCategorias.DisplayMember = "CategoryName"; //lo que ve usuario
+                this.cbCategorias.ValueMember = "CategoryID"; //valor que guarda el programa
+                this.cbCategorias.SelectedIndex = 0;
+                this.textBoxcategoryID.Text = string.Empty; 
+            }
             
         }
 
@@ -159,17 +201,32 @@ namespace northwing
                 if (!this.textBoxproductID.Text.Equals("") )
                 {
 
-                    nRows = productoController.eliminarProducto(this.textBoxproductID.Text);
-                    if (nRows > 0)
+                    string mensaje = "¿Seguro que deseas eliminar el producto?";
+                    string titulo = "Eliminar producto";
+                    MessageBoxButtons botones=MessageBoxButtons.YesNo;
+                    DialogResult result;
+                    result=MessageBox.Show(mensaje, titulo, botones);
+                    
+                    if (result == DialogResult.Yes)
                     {
-                        MessageBox.Show("Producto eliminado con éxito");
-                        borrarDatos();
-                        this.btAlta.Visible = false;
-                        this.btModificar.Visible = false;
-                        this.bteliminarproducto.Visible = false;
-                        this.btCancelar.Visible = false;
+                        
+                        nRows = productoController.eliminarProducto(this.textBoxproductID.Text);
+                        if (nRows > 0)
+                        {
+                            MessageBox.Show("Producto eliminado con éxito");
+                            borrarDatos();
+                            this.btañadirproducto.Visible = false;
+                            this.btModificar.Visible = false;
+                            this.bteliminarproducto.Visible = false;
+                            this.btCancelar.Visible = false;
+                        }
 
                     }
+                    else
+                    {
+                        MessageBox.Show("Gracias por confiar en nosotros");
+                    }
+                
                 }
             }
             catch (Exception ex)
@@ -186,17 +243,32 @@ namespace northwing
                 int nRows;
                 if(!this.textBoxproductID.Text.Equals(""))
                 {
-                    nRows = productoController.modificarProducto(this.textBoxproductID.Text,this.textBoxprecioxunidad.Text);
-                    if (nRows > 0)
-                    {
-                        MessageBox.Show("Precio producto modificado con éxito"); 
-                        borrarDatos();
-                        this.btAlta.Visible = false;
-                        this.btModificar.Visible = false;
-                        this.bteliminarproducto.Visible = false;
-                        this.btCancelar.Visible = false;
+                    string mensaje = "¿Seguro que deseas modificar el precio del producto?";
+                    string titulo = "Modificar Dato";
+                    MessageBoxButtons botones = MessageBoxButtons.YesNo;
+                    DialogResult result;
+                    result = MessageBox.Show(mensaje, titulo, botones);
 
+                    if (result == DialogResult.Yes)
+                    {
+                        nRows = productoController.modificarProducto(this.textBoxproductID.Text, this.textBoxprecioxunidad.Text);
+                        if (nRows > 0)
+                        {
+                            MessageBox.Show("Precio producto modificado con éxito");
+                            borrarDatos();
+                            this.btañadirproducto.Visible = false;
+                            this.btModificar.Visible = false;
+                            this.bteliminarproducto.Visible = false;
+                            this.btCancelar.Visible = false;
+
+                        }
                     }
+                    else
+                    {
+                        MessageBox.Show("Modifique sus datos cuando necesite");
+                    }
+
+                 
                 }
             }
             catch(Exception ex)
@@ -204,5 +276,13 @@ namespace northwing
                 MessageBox.Show(ex.Message);
             } 
         }
+
+        private void textBoxproductID_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            //te metes en diseño en el text box que necesites, pulsas rayo y seleccionas KEY PRESS
+            //pulsas dos veces al text box y en el código ponemos lo siguiente
+            //de esta manera el usuario no podrá introducir un valor no valido
+            //e.Handled = !char.IsDigit(e.KeyChar); 
+        } 
     }  
 }
